+++ runner/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@koa/cors": "^2.2.1",
+    "execa": "^0.8.0",
     "koa": "^2.3.0",
     "koa-body": "^2.5.0",
     "stream-to-string": "^1.1.0"
