--- runner/src/is-post.ts
@@ -1,9 +1,0 @@
-export default function() {
-  return function(ctx, next) {
-    if (ctx.method === 'POST') {
-      next()
-    } else {
-      ctx.status = 404
-    }
-  }
-}
